feat: Add MPS support and translate UI to English

+++ wan/models/motion_controller.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 
 
 def sinusoidal_embedding_1d(dim, position):
-    sinusoid = torch.outer(position.type(torch.float64), torch.pow(10000, -torch.arange(dim//2, dtype=torch.float64, device=position.device).div(dim//2)))
+    sinusoid = torch.outer(position.type(torch.float32), torch.pow(10000, -torch.arange(dim//2, dtype=torch.float32, device=position.device).div(dim//2)))
     x = torch.cat([torch.cos(sinusoid), torch.sin(sinusoid)], dim=1)
     return x.to(position.dtype)
